chore: move service type enum
