+++ src/scheduler/workers/processListsBeforeAndDuringStart/audit.ts
@@ -1,34 +1,11 @@
 import { prisma } from "shared/prisma";
-import { AuditEvent } from "@prisma/client";
-
-import {
-  AuditCreateInput, ListEventJsonData, ListItemEventJsonData
-} from "./types";
-
-/**
- * @deprecated
- * TODO: deprecate, this is handled by the history field
- */
-export function recordListItemEvent(
-  eventData: ListItemEventJsonData | ListEventJsonData,
-  auditEvent: AuditEvent,
-  type?: "user" | "list" | "listItem"
-) {
-  type = type ?? "listItem";
-  const data: AuditCreateInput = {
-    auditEvent,
-    type,
-    jsonData: { ...eventData },
-  };
-
-  return prisma.audit.create({ data });
-}
+import type { AuditEvent } from "@prisma/client";
 
 export async function findAuditEvents(
   annualReviewReference: string,
   auditEvent: AuditEvent,
   type?: "user" | "list" | "listItem",
-  itemId?: number,
+  itemId?: number
 ) {
   type = type ?? "listItem";
 
@@ -41,7 +18,8 @@ export async function findAuditEvents(
         path: ["annualReviewRef"],
         equals: annualReviewReference,
       },
-    }];
+    },
+  ];
   if (itemId) {
     andCondition.push({
       jsonData: {
