+++ src/scheduler/workers/unpublish/day/changeState/addAudit.ts
@@ -1,6 +1,6 @@
-import { prisma } from "shared/prisma";;
-import { AuditCreateInput, ListEventJsonData } from "server/models/types";
-import { AuditEvent } from "@prisma/client";
+import { prisma } from "shared/prisma";
+import type { AuditCreateInput, ListEventJsonData } from "shared/types";
+import type { AuditEvent } from "@prisma/client";
 
 export async function addAudit(eventData: ListEventJsonData, auditEvent: AuditEvent) {
   const data: AuditCreateInput = {
