+++ src/scheduler/workers/unpublish/day/changeState/resetAnnualReviewForList.ts
@@ -1,6 +1,6 @@
-import { prisma } from "shared/prisma";;
+import { prisma } from "shared/prisma";
 import { AuditEvent, List } from "@prisma/client";
-import { ListJsonData } from "server/models/types";
+import { ListJsonData } from "shared/types";
 import { addYears } from "date-fns";
 import { addAudit } from "./addAudit";
 import { Meta } from "../types";
