+++ src/scheduler/workers/unpublish/day/changeState/unpublishListItems.ts
@@ -1,5 +1,5 @@
-import { prisma } from "shared/prisma";;
-import { EVENTS } from "server/models/listItem/listItemEvent";
+import { prisma } from "shared/prisma";
+import { EVENTS } from "shared/listItemEvent";
 
 export async function unpublishListItems(ids: number[], reference?: string) {
   return ids.map(async (listItemId: number) => {
