+++ src/scheduler/workers/unpublish/day/dayReminderPersonalisation.ts
@@ -1,4 +1,4 @@
-import { ServiceType } from "server/models/types";
+import { ServiceType } from "shared/types";
 import { Meta } from "./types";
 import { List, ListItem } from "@prisma/client";
 import { ListItemJsonData } from "server/models/listItem/providers/deserialisers/types";
