+++ src/scheduler/workers/unpublish/weekly/weeklyReminderPersonalisation.ts
@@ -1,4 +1,4 @@
-import { ServiceType } from "server/models/types";
+import type { ServiceType } from "shared/types";
 import { ListItem } from "@prisma/client";
 import { ListItemJsonData } from "server/models/listItem/providers/deserialisers/types";
 import { createAnnualReviewProviderUrl } from "scheduler/workers/createAnnualReviewProviderUrl";
