+++ src/server/components/formRunner/lawyers.ts
@@ -1,4 +1,5 @@
-import { LawyerListItemGetObject, ServiceType } from "server/models/types";
+import { LawyerListItemGetObject } from "server/models/types";
+import { ServiceType } from "shared/types";
 import { Field, Question } from "server/components/formRunner/types";
 import { parseJsonFormData } from "server/components/formRunner/helpers";
 import { get } from "lodash";
