+++ src/server/components/formRunner/types.ts
@@ -1,4 +1,4 @@
-import { ServiceType } from "server/models/types";
+import type { ServiceType } from "shared/types";
 
 export interface Component {
   name: string;
