+++ src/server/components/lists/constants.ts
@@ -1,5 +1,5 @@
 import { countriesList } from "server/services/metadata";
-import { ServiceType } from "server/models/types";
+import { ServiceType } from "shared/types";
 import { listsRoutes } from "./routes";
 import { dashboardRoutes } from "server/components/dashboard";
