+++ src/server/components/lists/controllers/ingest/ingestPutController.ts
@@ -2,10 +2,10 @@ import { Request, Response } from "express";
 import { formRunnerPostRequestSchema } from "server/components/formRunner";
 import { logger } from "server/services/logger";
 import { prisma } from "shared/prisma";;
-import { recordListItemEvent } from "server/models/audit";
+import { recordListItemEvent } from "shared/audit";
 import { AuditEvent, Prisma, Status } from "@prisma/client";
 import { DeserialisedWebhookData } from "server/models/listItem/providers/deserialisers/types";
-import { ListJsonData, ServiceType } from "server/models/types";
+import { ListJsonData, ServiceType } from "shared/types";
 import { deserialise } from "server/models/listItem/listItemCreateInputFromWebhook";
 import { getServiceTypeName } from "server/components/lists/helpers";
 import { EVENTS } from "server/models/listItem/listItemEvent";
