+++ src/server/components/lists/questionnaire.ts
@@ -1,6 +1,6 @@
 import type { Request } from "express";
 import { startCase, kebabCase } from "lodash";
-import { ServiceType } from "server/models/types";
+import { ServiceType } from "shared/types";
 import { parseListValues, getServiceLabel, getAllRequestParams } from "./helpers";
 import type { QuestionName, Question, QuestionData } from "./types";
 import {
