+++ src/server/components/sitemap/controllers.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import { ServiceType } from "server/models/types";
+import { ServiceType } from "shared/types";
 import { countriesList } from "server/services/metadata";
 import { listsRoutes, getServiceLabel } from "server/components/lists";
 import { pageTitles } from "server/components/dashboard/helpers";
