+++ src/server/models/helpers.ts
@@ -1,8 +1,9 @@
 import { isNumber, isArray } from "lodash";
 import { getDbPool } from "./db/database";
-import { CountryName, ServiceType } from "server/models/types";
+import { CountryName } from "server/models/types";
+import { ServiceType } from "shared/types";
 import { findListByCountryAndType } from "server/models/list";
-import { prisma } from "shared/prisma";;
+import { prisma } from "shared/prisma";
 import { logger } from "server/services/logger";
 
 /**
