+++ src/server/models/listItem/providers/FuneralDirectors.ts
@@ -1,7 +1,8 @@
-import { FuneralDirectorListItemGetObject, ServiceType } from "server/models/types";
+import { FuneralDirectorListItemGetObject } from "server/models/types";
+import { ServiceType } from "shared/types";
 import { getPlaceGeoPoint } from "./../geoHelpers";
 import { logger } from "server/services/logger";
-import { prisma } from "shared/prisma";;
+import { prisma } from "shared/prisma";
 import { fetchPublishedListItemQuery } from "server/models/listItem/providers/helpers";
 
 export async function findPublishedFuneralDirectorsPerCountry(props: {
