+++ src/server/models/listItem/providers/Lawyers.ts
@@ -1,4 +1,5 @@
-import { LawyerListItemGetObject, ServiceType } from "server/models/types";
+import { LawyerListItemGetObject } from "server/models/types";
+import { ServiceType } from "shared/types";
 import { getPlaceGeoPoint } from "./../geoHelpers";
 import { logger } from "server/services/logger";
 import { prisma } from "shared/prisma";;
